\name{aplMeta}
\alias{aplMeta}
\title{ alignedPeakListMetaData generation}
\description{
  Function alpMeta uses tofListMetaData to generate alignedPeakListMetaData
}
\usage{
aplMeta(tofListMetaData)
}
\arguments{
  \item{tofListMetaData}{ -- meta-data array for tofList }
}
\author{Maureen Tracy, William and Mary Research Institute, mbtrac@wm.edu }