\name{circShift}
\alias{circShift}
\title{ Circular shift of data record }
\description{
  Performs a circular shift of TOF record, e.g., for phase shift correction
}
\usage{
circShift(dataV, shiftN)
}
\arguments{
  \item{dataV}{ -- data vector}
  \item{shiftN}{ -- points to shift }
}
\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
