\name{coefTrunc}
\alias{coefTrunc}
\title{Wavelet truncation}
\description{
  Target wavelet amplitude truncation to instrumental precision
}
\usage{
coefTrunc(fcoef)
}
\arguments{
  \item{fcoef}{ -- wavelet coefficients}
}
\references{ Malyarenko D., et.al., Rapid Commun. Mass Spec., 2006, 20 (11), p.1661  }
\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }