\name{gausHalf}
\alias{gausHalf}
\title{ Left-half Gaussian wavelet}
\description{
  Left-half Gaussian wavelet calculation
}
\usage{
gausHalf(gt, trnct)
}
\arguments{
  \item{gt}{ -- half-width in time points }
  \item{trnct}{ -- wavelet amplitude truncation threshold}
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
