\name{gausRHalf}
\alias{gausRHalf}
\title{ Right-half Gaussian wavelet}
\description{
  Right-half Gaussian wavelet calculation
}
\usage{
gausRHalf(gt, trnct)
}
\arguments{
  \item{gt}{ -- half-width in time points }
  \item{trnct}{ -- wavelet amplitude truncation threshold}
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
