\name{lorHalf}
\alias{lorHalf}
\title{ Right-half Lorentzian wavelet}
\description{
  Right-half Lorentzian wavelet calculation
}
\usage{
lorHalf(gt, trnct)
}
\arguments{
  \item{gt}{ -- half-width in time points }
  \item{trnct}{ -- wavelet amplitude truncation threshold}
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
