\name{lorLHalf}
\alias{lorLHalf}
\title{ Left-half Lorentzian wavelet}
\description{
  Left-half Lorentzian wavelet calculation
}
\usage{
lorLHalf(gt, trnct)
}
\arguments{
  \item{gt}{ -- half-width in time points }
  \item{trnct}{ -- wavelet amplitude truncation threshold}
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
