\name{matchedFilter}
\alias{matchedFilter}
\title{ Matched filter}
\description{
  Produces matched filter noise smoothing for the input signal
}
\usage{
matchedFilter(s, h)
}
\arguments{
  \item{s}{ -- matrix with observed sampled discrete signals }
  \item{h}{ -- noiseless signal wavelet (best fit) }
}

