\name{matchFiltCoefGG}
\alias{matchFiltCoefGG}
\title{ Matched filter coefficients for Gaussian signal }
\description{
  Calcualtes matched filter coefficients for Gaussian signal wavelet model
}
\usage{
matchFiltCoefGG(inhw, las, trct)
}
\arguments{
  \item{inhw}{ -- left half-width of an input signal }
  \item{las}{ -- right half-width asymmetry (can be negative) }
  \item{trct}{ -- target wavelet amplitude truncation threshold}
}

