\name{matchFiltCoefGL}
\alias{matchFiltCoefGL}
\title{ Matched filter coefficients for Gaussian-Lorentzian signal }
\description{
  Calcualtes matched filter coefficients for Gaussian-Lorentzian signal wavelet model
}
\usage{
matchFiltCoefGL(inhw, las, trct)
}
\arguments{
  \item{inhw}{ -- left half-width of an input signal }
  \item{las}{ -- right half-width asymmetry (can be negative) }
  \item{trct}{ -- target wavelet amplitude truncation threshold}
}