\name{matchFiltCoefLL}
\alias{matchFiltCoefLL}
\title{ Matched filter coefficients for Lorentzian signal }
\description{
  Calcualtes matched filter coefficients for Lorentzian signal wavelet model
}
\usage{
matchFiltCoefLL(inhw, las, trct)
}
\arguments{
  \item{inhw}{ -- left half-width of an input signal }
  \item{las}{ -- right half-width asymmetry (can be negative) }
  \item{trct}{ -- target wavelet amplitude truncation threshold}
}