\name{mavSmoothing}
\alias{mavSmoothing}
\title{ Moving average filter}
\description{
  Applies moving average filter to smooth input
}
\usage{
mavSmoothing(inputSig, pointWindowLength)
}
\arguments{
  \item{inputSig}{ -- input spectrum signal }
  \item{pointWindowLength}{ -- filter window width }
}

\author{ Maureen Tracy, William and Mary Research Institute, mbtrac@wm.edu }

