\name{noiseEstmtr2}
\alias{noiseEstmtr2}
\title{ Gaussian noise STD-estimator }
\description{
  Noise estimate assuming Gaussian (random) noise and constant peak-width
}
\usage{
noiseEstmtr2(spect)
}
\arguments{
  \item{spect}{ -- spectrum record with the noise and constant peak-width}
}
\author{ William E Cooke, Professor, Physics, College of William and Mary, wecook@wm.edu }

