\name{optFilt}
\alias{optFilt}
\title{ Optimal linear filter }
\description{
  Noise smoothing without signal broadening
}
\usage{
optFilt(inSig, wfc)
}
\arguments{
  \item{inSig}{ -- input TOF signal }
  \item{wfc}{ -- filter coefficients }
}
\references{ Malyarenko D., et.al., Rapid Commun. Mass Spec., 2006, 20 (11), p.1661  }
\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }

