\name{OutputForML}
\alias{OutputForML}
\title{ Output for Matlab  }
\description{
  Reformat R-structures into Matlab output
}
\usage{
OutputForML(object2save, choice)
}
\arguments{
  \item{object2save}{ -- object to save }
  \item{choice}{ -- structure choice: "ISP"- inputSpecList,"TRS"- tofResampled,"PL"- peakList,"APL"-alignedPeakList }
}

\author{ Maureen Tracy, William and Mary Research Institute, mbtrac@wm.edu  }
