\name{pivPeak}
\alias{pivPeak}
\title{ Pivot peak detection }
\description{
  Finds time position of the pivot peaks above threshold
}
\usage{
pivPeak(refsig, thres, phw)
}
\arguments{
  \item{refsig}{ -- reference signal normalized to its maximum }
  \item{thres}{ -- intensity threshold for pivot peaks in respect to maximum }
  \item{phw}{ -- peak half-width in the pivot detection range }
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu }
