\name{plMeta}
\alias{plMeta}
\title{ Peak list meta-data maker}
\description{
  Compiles peak list meta data from TOF list meta-data
}
\usage{
plMeta(tofListMetaData)
}
\arguments{
  \item{tofListMetaData}{ -- TOF list meta-data }
}
\author{ Maureen Tracy, William and Mary Research Institute, mbtrac@wm.edu }
