\name{trgFilt}
\alias{trgFilt}
\title{ Target filter coefficients }
\description{
  Solves linear system for target filter coefficients
}
\usage{
trgFilt(sWvlt, nu, dWvlt)
}
\arguments{
  \item{sWvlt}{ -- signal wavelet }
  \item{nu}{ -- noise power ratio to signal power }
  \item{dWvlt}{ --desired signal wavelet shifted in respect to input }
}

\author{ Dariya Malyarenko, William and Mary Research Institute, dimaly@wm.edu}
