\name{userError}
\alias{userError}
\title{Stops execution with user defined message}
\description{
  Output custom message with error
}
\usage{
userError(msg)
}
\arguments{
  \item{msg}{ -- text of the error message }
}
\author{ Maureen Tracy, William and Mary Research Institute, mbtrac@wm.edu }
